Verify if data is encoded before publishing
As publishData() doesn't verify if 'data.value' contains characters
forbidden by Fiware-Orion (<, >, ", ', (, ), =, ;), requests sent
to Orion with Base64 values (supported by KNoT) will result in a
'400 Bad Request' response.

As a result, Base64 values with '=' will not be updated and will
remain as is, or with an empty string (" ") as its value, in Orion.
Add isBase64() to verify if the 'data.value' is Base64 encoded and
update publishData to decode Base64 values before publishing them.